Delete 'commont.js'
