+++ src/utils/util.js
@@ -58,6 +58,20 @@ const getRandomeDate = (maxPassedDays) => {
 
 const isEscape = (code) => code === 'Escape';
 
+const updateItem = (items, update) => {
+  const index = items.findIndex((item) => item.id === update.id);
+
+  if (index === -1) {
+    return items;
+  }
+
+  return [
+    ...items.slice(0, index),
+    update,
+    ...items.slice(index + 1),
+  ];
+};
+
 export {
   getRandomInteger,
   getRandomFloat,
@@ -70,5 +84,6 @@ export {
   getMockText,
   getRandomBoolean,
   getRandomeDate,
-  isEscape
+  isEscape,
+  updateItem
 };
